+++ libdfu/dfu-image-private.h
@@ -27,7 +27,7 @@
 G_BEGIN_DECLS
 
 DfuImage	*dfu_image_from_dfuse		(const guint8	*data,
-						 gsize		 length,
+						 guint32	 length,
 						 guint32	*consumed,
 						 GError		**error);
 GBytes		*dfu_image_to_dfuse		(DfuImage	*image);
